Revert "Merge pull request #37 from quinn-dougherty/refactor"
This reverts commit b9f3c88a0719ae6a051a4fb2db48e697da4add17, reversing
changes made to 6dac2d73fb57a93785b8e2250a0a7c6ba277a669.
