+++ Dockerfile
@@ -12,4 +12,4 @@ RUN chmod u+x setup.sh && PORT=8000 ./setup.sh
 
 # expanding shell variables in CMD is tricky, see
 # https://stackoverflow.com/questions/23071214/use-environment-variables-in-cmd
-CMD ["streamlit", "run", "--server.port", "8000", "src/app.py"]
+CMD ["streamlit", "run", "--server.port", "8000", "app.py"]
