+++ Procfile
@@ -1,1 +1,1 @@
-web: sh setup.sh && streamlit run src/app.py
+web: sh setup.sh && streamlit run app.py
