--- src/constants.py
@@ -1,13 +1,0 @@
-#!/usr/bin/env python
-
-
-DELAWARE = 564696
-CHESTER = 519293
-MONTGOMERY = 826075
-BUCKS = 628341
-PHILLY = 1581000
-S_DEFAULT = DELAWARE + CHESTER + MONTGOMERY + BUCKS + PHILLY
-KNOWN_INFECTIONS = 31
-
-
-RECOVERY_DAYS = 14.0
